pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "pixrr"
version = "0.1.0"
description = "A lightweight image processing toolkit for Python"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Hrishikesh Tiwari", email = "tiwarihrishikesh686@gmail.com" }
]

requires-python = ">=3.9"

dependencies = [
    "numpy>=2.3.5",
    "scipy>=1.16.3",
    "matplotlib>=3.10.7",
    "pillow>=12.0.0",
    "numba>=0.63.1"
]

keywords = ["image-processing", "computer-vision", "education", "filters", "numpy"]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Intended Audience :: Education",
]


[project.urls]
"Homepage" = "https://github.com/Hrishi11572/pixrr"
"Documentation" = "https://github.com/Hrishi11572/pixrr/tree/main/docs"

